Changed the way coordinate systems work in regions. All regions now use a topleft (0, 0) based coordinate system internally, with m_inner_bounds the definition of this. Regions also store their own rectangle in the parent coordinate system as m_outer_bounds. Conversion is done when calling parent need_redraw() or child redraw() methods.
Redraw() has a different signature, passing the dirty area in local
coordinates along with a reference to the toplevel window and the
coordinates of local (0, 0) in toplevel window space.

active_renderer::redraw() has become a true virtual function, and
the old default functionality of it has moved to a new class
none_active_renderer, so you'll get compiler errors if you don't
override it after the signature change.