We need to match type of objects to be passed to C++ exactly, otherwise we will not get any extended behaviour added in (Python) subclasses of the C++ class.
Therefore, started on creation of C++->Python wrapper objects on demand,
i.e. when a Python object is passed to C++.

This revealed a couple of problems with not all vritual methods of the
(C++) baseclass being suitably overridden, due to (a) "const" parameter
problems and (b) methods not being representable in Python.

Solved first case by getting "const" right. Started on solving the second
by allowing the insterion of dummy methods into the wrapped class definition.