Revamped drop logic: now that we have early dropping (both before and after the decoder, which is the actually expensive operation) never drop frames late (i.e. in the video renderer). In stead, even if the frame is late, display it anyways and tell the decoder datasource to skip frames up to "now" (which will happen early, again).
This results in a serious increase in fps, at the expense of lipsync, but we probably
shouldn't worry overmuch about that anyway on devices underpowered for the
video stream at hand.

Also, initial tests seem to indicate that another twofold increase in fps can be gotten
by fixing the multiple copying in the backend renderer.