Fix for #3004655: audio is not lipsync.
The core problem is that our current audio clock sync implementation is far
from complete, and it specifically breaks if there are multiple audio
renderers active at the same time.

The stopgap is that in that case one of the audio renderers (usually the
first one started) "wins" and syncs the system clock to its audio
clock.