Preparing for a better startup scheme (where we initialize only the things we need).
Step 1 - get rid of most properties in favor of a method-based
scheme for the AmbulantViewController. May have broken some things,
to be fixed later.

Additionally, we record the fact that we have finished viewing the
whole presentation (so we don't restart on next application launch
at the very last beat of the app).