Fix for #351842 and a secon bug that showed up while fixing it.
3531842 is about the SDL audio renderer terminating early if it is
paused during an audio node that has a clip_end. The problem was
a logic error in ffmpegi_decoder_datasource::_clip_end(), which
looked at the timestamp of the incoming data in stead of the timestamp
of the outgoing (client-consumed) data.

Fixing this revealed a new bug, which we have also seen before:
sometimes the audio of a multiplexed A/V file can be played twice.
The problem turns out to be that if the video is seeked but the
audio is not there is no flush, therefore we get double audio
data. Fixed with a hack: the video renderer calls preroll
with howmuch==-1 to force the flush and the seek.